capture exceptions

+++ src/process_benchmarks/smt2eq_ostrich.py
@@ -6,21 +6,36 @@ import os
 import shutil
 import subprocess
 def main():
-    smt_file_folder=bench_folder+"/kaluzaWoorpje/smt2"
-    eq_file_folder=bench_folder+"/kaluzaWoorpje/eq"
+    benchmark="kaluza/kaluzaSmallSatExtracted"
+    smt_file_folder=bench_folder+f"/{benchmark}/smt2"
+    eq_file_folder=bench_folder+f"/{benchmark}/eq"
+    exception_file_folder=bench_folder+f"/{benchmark}/exceptions"
     ostrich_output_file=bench_folder+"/temp/output.eq"
     solver="ostrich_export"
 
     update_ostrich()
 
+    for answer_file in glob.glob(smt_file_folder+"/*.answer"):
+        os.remove(answer_file)
+
+    exception_list=[]
     for smt_file in glob.glob(smt_file_folder+"/*.smt2"):
         if os.path.exists(ostrich_output_file):
             os.remove(ostrich_output_file)
 
+
         smt_file_path=os.path.join(smt_file_folder,smt_file)
         result_dict = run_on_one_problem(file_path=smt_file_path, parameters_list=["-timeout=0"], solver=solver)
         file_name=strip_file_name_suffix(os.path.basename(smt_file_path))
-        shutil.copy(ostrich_output_file,eq_file_folder+f"/{file_name}.eq")
+        if os.path.exists(ostrich_output_file):
+            shutil.copy(ostrich_output_file,eq_file_folder+f"/{file_name}.eq")
+        else:
+            exception_list.append(smt_file_path)
+            shutil.copy(smt_file,exception_file_folder)
+    print("exception_list:",exception_list)
+
+
+
 
 @time_it
 def update_ostrich():
